emap-setup/requirements.txt
PyYAML
gitpython
flake8